fix: Resolve build errors and clippy warnings
- Add .cargo/config.toml to use Clang instead of GNAT toolchain's C++
  compiler, fixing RocksDB compilation failures
- Add module-level clippy allows for common style lints in orbit-server
- Fix approximate PI value warnings in test files with targeted allows
- Rename HyperLogLog to_string/from_string to encode_to_string/decode_from_string
  to avoid shadowing Display trait
- Update manual range checks to use .contains() method
- Simplify redundant closures to function references
- Change assert_eq!(x, true) to assert!(x) in bolt protocol tests
- Replace &[val.clone()] with std::slice::from_ref(&val)
- Restructure MutexGuard usage to avoid holding across await points
- Fix filter_map that always returns Some to use map instead
- Remove unnecessary borrows, format! calls, and casts

+++ orbit/server/benches/simd_benchmarks.rs
@@ -597,7 +597,7 @@ fn bench_operators(c: &mut Criterion) {
     // I should follow that pattern for consistency and isolation.
 
     for size in [1000, 100_000].iter() {
-        let values: Vec<i32> = (0..*size).map(|i| (i % 100) as i32).collect();
+        let values: Vec<i32> = (0..*size).map(|i| i % 100).collect();
         let target = 50;
 
         // Scalar
